Add api changes
